1. Setup VKSdk Delegate in View controller 2. Share dialog with bug auth in iOS 9.0

+++ Randomize Me/Models/Responses/MSRandomResponse.m
@@ -108,7 +108,7 @@
     
     NSString *individualInformation = @"Individual information of generation:";
     NSString *completionTime = [NSString stringWithFormat:@"Completion time (UTC+0): %@", [self makeStringComplitionTime]];
-    NSString *serialNumber = [NSString stringWithFormat:@"Serial Number: %d", self.serialNumber];
+    NSString *serialNumber = [NSString stringWithFormat:@"Serial Number: %ld", (long)self.serialNumber];
     NSString *signature = [NSString stringWithFormat:@"Signature: %@", self.signature];
     
     NSString *result = [NSString stringWithFormat:@"%@\r\r%@\r%@\r\r%@\r%@\r%@\r%@\r%@\r\r%@\r%@\r%@\r%@", resultName, forResult, resultData, parametrs, numberOfIntegers, minValue, maxValue, replacement, individualInformation, completionTime, serialNumber, signature];
